// contracts/typescript/models/MutationReceipt.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface MutationReceipt
 */
export interface MutationReceipt {
    /**
     *
     */
    calendarVersion: number;
    /**
     *
     */
    contextId: string;
    /**
     *
     */
    eventId: string;
    /**
     *
     */
    version: number;
}

/**
 * Check if a given object implements the MutationReceipt interface.
 */
export function instanceOfMutationReceipt(value: object): value is MutationReceipt {
    if (!('calendarVersion' in value) || value['calendarVersion'] === undefined) return false;
    if (!('contextId' in value) || value['contextId'] === undefined) return false;
    if (!('eventId' in value) || value['eventId'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    return true;
}

export function MutationReceiptFromJSON(json: any): MutationReceipt {
    return MutationReceiptFromJSONTyped(json, false);
}

export function MutationReceiptFromJSONTyped(json: any, ignoreDiscriminator: boolean): MutationReceipt {
    if (json == null) {
        return json;
    }
    return {

        'calendarVersion': json['calendarVersion'],
        'contextId': json['contextId'],
        'eventId': json['eventId'],
        'version': json['version'],
    };
}

export function MutationReceiptToJSON(json: any): MutationReceipt {
    return MutationReceiptToJSONTyped(json, false);
}

export function MutationReceiptToJSONTyped(value?: MutationReceipt | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'calendarVersion': value['calendarVersion'],
        'contextId': value['contextId'],
        'eventId': value['eventId'],
        'version': value['version'],
    };
}