// contracts/typescript/models/OnsiteView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues, parseDate, parseDateTime, serializeDate, serializeDateTime } from '../runtime';
import type { OnsiteState } from './OnsiteState';
import {
    OnsiteStateFromJSON,
    OnsiteStateFromJSONTyped,
    OnsiteStateToJSON,
    OnsiteStateToJSONTyped,
} from './OnsiteState';

/**
 *
 * @export
 * @interface OnsiteView
 */
export interface OnsiteView {
    /**
     *
     */
    employeeId: string;
    /**
     *
     */
    from: Date;
    /**
     *
     */
    id: string;
    /**
     *
     */
    location: string;
    /**
     *
     */
    readAt: Date | null;
    /**
     *
     */
    reason: string;
    /**
     *
     */
    reference: string;
    /**
     *
     */
    revision: number;
    /**
     *
     */
    state: OnsiteState;
    /**
     *
     */
    to: Date;
    /**
     *
     */
    version: number;
}



/**
 * Check if a given object implements the OnsiteView interface.
 */
export function instanceOfOnsiteView(value: object): value is OnsiteView {
    if (!('employeeId' in value) || value['employeeId'] === undefined) return false;
    if (!('from' in value) || value['from'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('location' in value) || value['location'] === undefined) return false;
    if (!('readAt' in value) || value['readAt'] === undefined) return false;
    if (!('reason' in value) || value['reason'] === undefined) return false;
    if (!('reference' in value) || value['reference'] === undefined) return false;
    if (!('revision' in value) || value['revision'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    if (!('to' in value) || value['to'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    return true;
}

export function OnsiteViewFromJSON(json: any): OnsiteView {
    return OnsiteViewFromJSONTyped(json, false);
}

export function OnsiteViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): OnsiteView {
    if (json == null) {
        return json;
    }
    return {

        'employeeId': json['employeeId'],
        'from': (json['from'] == null ? json['from'] : parseDate(json['from'])),
        'id': json['id'],
        'location': json['location'],
        'readAt': (json['readAt'] == null ? null : parseDateTime(json['readAt'])),
        'reason': json['reason'],
        'reference': json['reference'],
        'revision': json['revision'],
        'state': OnsiteStateFromJSON(json['state']),
        'to': (json['to'] == null ? json['to'] : parseDate(json['to'])),
        'version': json['version'],
    };
}

export function OnsiteViewToJSON(json: any): OnsiteView {
    return OnsiteViewToJSONTyped(json, false);
}

export function OnsiteViewToJSONTyped(value?: OnsiteView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'employeeId': value['employeeId'],
        'from': value['from'] == null ? value['from'] : serializeDate(value['from']),
        'id': value['id'],
        'location': value['location'],
        'readAt': value['readAt'] == null ? value['readAt'] : serializeDateTime(value['readAt']),
        'reason': value['reason'],
        'reference': value['reference'],
        'revision': value['revision'],
        'state': OnsiteStateToJSON(value['state']),
        'to': value['to'] == null ? value['to'] : serializeDate(value['to']),
        'version': value['version'],
    };
}